Visual Aids
I added some images just to show what the original drawings looked like. Of course, I did not draw them the entire way through, but I just put a few of them online just to give users an idea of how the brainchild looked like in its embryonic stage. And that is not counting those times when they were just squares drawn on the fly!