Mute Music Button
Alright, so...some people might not want to hear 8-bit music the whole time they are playing. Therefore, I added a button so that they can turn it off if they want to. I have yet to see if it affects the gameplay at all. However, based on what I'm seeing, they are just little bits that do not cause that many problems to the operations of the program. Of course, I might be wrong, but we will see.