A New Experiment
I decided to see what would happen if I had a way to mute the music when the questions were being asked. I figured that doing it this way would work rather than just muting the music the whole way through. If this does not work, however, I will have to preserve other versions of the original so that I can revert to the original code when I have time.