New Changes
After working on a research project, my brain felt like it was on fire, so I decided to take that to my own advantage and do some coding.

I made several changes to the user interface of the instructions page. I centered the table (turns out it just involves having left and right margins position themselves automatically), added some responsive padding to the paragraph tags and most of the headers, and aligned the lists with the paragraphs so that they seemed together.

This was a nice brain teaser. I needed that.

Also, it's been a while since I've done work using CSS. I believe it is important for me to know how to work with this language.