More Code Changes
Finally got the question shuffling function to work. For this reason, I've hardwired the game to reshuffle the questions all over again. This is basically like shuffling cards. When a game is over, you get a new set. Might have the same cards as before. only in a different order.

It's nice to get back to writing code. Wish I did this more often.