Even Easey-er 2 Play!
Might be a funny title, but that was the goal for this code-editing session.

Today, I added a PLAY button for mobile users to see when the device is held in portrait mode. To make sure they saw it, I made it strobe so that people knew what to click.

Of course, this required new HTML elements, new CSS attributes, and new JavaScript commands. In other words, new code. (In fact, coding has actually gotten easier in a sense. I've gotten so used to it that now I just look for pointers and then fill in the rest of the blanks.)

Anyhow, that's enough for now.