Bug Fixing
I had no idea that there was still a bug. It was a rather simple one: the high score is cleared once the game is played again, which is not what I wanted. For this reason, I've written some code that explicitly tells the computer to keep the high score where it was last at, even if the game restarts.