New Logo!
When I first started the company, I did not have a logo. Now I do!

Had to write new code in order to make incorporate the logo into the current design. Looks as if I will have to give this site a major redesign. Perhaps a nice site with a dark background? Who knows.