Just for Arcade Machines
I decided I did not want to do an analysis of the entire population since it would take a lot of time, and would require sending information to a form. For this reason, I won't be doing that. Since I will be displaying the video game using a video game controller, I decided to just do the analysis locally.