Fixing Darn Bugs
Had to fix a bug that kept on freezing the game. It had to do with the eyes of the rear Israelite. It was causing such a hiccup that I had to reset the computer and rewrite values a couple of times. For this reason, I had to put the information in a try-catch statement.