More Bug Fixes
For some reason, it still did not fix the whole issue. Hopefully this fixes most of the problems revolving around Google Pixel phones. Obviously, in the future I will have to port this. That's the problem. Porting will be a bit of a terror, especially if it is in a language that I don't know. I would have to learn to use C# and utilize the tools in Unity to create a successful port.