Search Engine Optimization Work
Most of the work I did here was to help optimize the search engine. I was trying to help the search engine as well as the generative artificial intelligence crawlers to find my website and put it in the AI summaries. For some reason, Google cannot seem to find this company. For this reason, I decided to write some code that would help it see the company and its information. In other words, I wrote more information. That's really what artificial intelligence models want. More information.

We'll see how well this website fares with the search engine as well as every single LLM that exists.